Added basic support for default column values
Implemented a new property annotation ("!DefaultValue") to specify a
default column value used to initially fill a POST form. The filling is
done during the to() method of the form.

Updated the HTML template and jQuery stuff on the recess tools to allow
the import of the default values from the database table structure
(tested only on MySQL) and added missing code generation to ModelGen.

Added a new test suite for ModelForm: Model::finalClassDescriptor() has
been changed to allow the test to run without database connection.

The major issue is the datetime management: actually any string value
is fed to strtotime(), so datetimes could be generated "dynamically" to
to prefill the form inputs. I think there is no sane way to convert
this value from SQL to PHP (and vice versa) but maybe a
strreplace('CURRENT_TIMESTAMP', 'NOW'...) will cover the most common
cases... the issue is still open to further improvements.