Differentiate between "id" and "name" attributes
The "id" and "name" attributes of form elements follow different rules:
the square brackets (used by default to generate array values in PHP)
are valids as "name" values but not as "id".

Added the FormInput::$id property (got by substituting '[]' with '-')
and updated the form rendering to use this new property.