Extending Form to support multiple submit inputs
Allow the $method property of a form to be specified as an associative
array (Http::method => label). If the method is not an array, the class
behaves like before, retaining backward compatibility.

To be able to manage multiple submit forms in IE6 (this piece of shit is
still widely used), some workaround has been implemented. The <button>
tag is not supported (all the <button> values are submitted, making
impossible to know which button was pressed) and any client side crap is
not an option. The only fairly supported tag is <input type="submit">.

It would be too easy to just use:

<input type="submit" name="_METHOD" value="put">Update</input>

This is the <button> way, in the input tag the value IS the label and
there is nothing can be done. The freedom to set a custom label is an
high priority, especially when you are not english, so I opted for:

<input type="submit" name="_METHOD[put]" value="Update" />

Embedding the HTTP method in the name leaves the value free to be set to
whatever. The implementation on the server side is quite trivial and can
be found inside DefaultPolicy::getHttpMethodFromPost().